// src/types/mock-data.type.ts
type MockData = {
  users: {
    names: string[];
    emails: string[];
    avatars: string[];
    passwords: string[];
  }
  films: {
    titles: string[];
    descriptions: string[];
    genres: string[];
    previewVideoLinks: string[];
    videoLinks: string[];
    starring: string[];
    directors: string[];
    posterImages: string[];
    backgroundImages: string[];
    backgroundColors: string[];
    comments: string[];
  }
}

export { type MockData };